server/storage_service/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
mongoengine==0.29.1
dnspython==2.7.0